+++ lithops/serverless/backends/ibm_cf/ibm_cf.py
@@ -110,7 +110,7 @@ class IBMCloudFunctionsBackend:
     def _delete_function_handler_zip(self):
         os.remove(ibmcf_config.FH_ZIP_LOCATION)
 
-    def build_runtime(self, docker_image_name, dockerfile):
+    def build_runtime(self, docker_image_name, dockerfile, extra_args=[]):
         """
         Builds a new runtime from a Docker file and pushes it to the Docker hub
         """
@@ -118,9 +118,14 @@ class IBMCloudFunctionsBackend:
         logger.info('Docker image name: {}'.format(docker_image_name))
 
         if dockerfile:
-            cmd = 'docker build -t {} -f {} .'.format(docker_image_name, dockerfile)
+            cmd = '{} build -t {} -f {} . '.format(ibmcf_config.DOCKER_PATH, docker_image_name, dockerfile)
         else:
-            cmd = 'docker build -t {} .'.format(docker_image_name)
+            cmd = '{} build -t {} . '.format(ibmcf_config.DOCKER_PATH, docker_image_name)
+
+        cmd = cmd+' '.join(extra_args)
+
+        if logger.getEffectiveLevel() != logging.DEBUG:
+            cmd = cmd + " >{} 2>&1".format(os.devnull)
 
         res = os.system(cmd)
         if res != 0: